Modificaciones cosmeticas
Se realizan modificaciones de presentación y limpieza
