+++ app/Models/TextManager/TextProcessor.php
@@ -5,15 +5,22 @@ use App\Models\Entities\UserCase;
 use App\Models\Entities\Operation;
 use App\Models\Utils\CustomGravilityException;
 use App\Models\TextManager\TextMapper;
+use Exception;
 
 class TextProcessor {
 
+	/**
+	 * Metodo encargado de procesar cada una de las lineas del texto ingresado por el usuario
+	 *
+	 * @param  String   $inputText  Texto ingresado por el usuario
+	 * @return Input 	Objeto resultante del texto procesado
+	 */ 
 	public function ValidateText($inputText){		
 		$input = new Input();
 		$input->userText = $inputText;
 		try{			
 			$lines = explode("\n", $inputText);		
-			$_testCases = array_shift($lines);
+			$_testCases = intval(array_shift($lines));
 			$input->save();
 			if($_testCases < 1 || $_testCases > 50){
 				throw new CustomGravilityException("Número de casos de pruebas no aceptados");		
@@ -38,7 +45,7 @@ class TextProcessor {
 			$input->passed = false;
 		}
 		catch(Exception $ex){
-			$input->outMessage = "Error inesperado - ".$ex->getMessage();
+			$input->outMessage = "TextProcessor Error inesperado - ".$ex->getMessage();
 			$input->passed = false;
 		}
 		$input->save();
