+++ tests/TextProcessorTest.php
@@ -1,9 +1,6 @@
 <?php
 
 use App\Models\TextManager\TextProcessor;
-use App\Models\Entities\Input;
-use App\Models\Entities\Operation;
-use App\Models\Entities\UserCase;
 
 class TextProcessorTest extends TestCase
 {
@@ -13,11 +10,16 @@ class TextProcessorTest extends TestCase
 		$this->assertInstanceOf('App\Models\Entities\Input',$processor->ValidateText($text));
 	}
 
-	public function test_Errors(){
+	public function test_WrongLinePassedFalse(){
 		$text = "0\n4 2\nUPDATE 2 2 2 4\nQUERY 1 1 1 3 3 3";
 		$processor = new TextProcessor();
-		$this->setExpectedException('\App\Models\Utils\CustomGravilityException', 'Número de casos de pruebas no aceptados');
-		$processor->ValidateText($text);
+		$this->assertFalse($processor->ValidateText($text)->passed);
+		$text = "0\n4 \nUPDATE 2 2 2 4\nQUERY 1 1 1 3 3 3";
+		$processor = new TextProcessor();
+		$this->assertFalse($processor->ValidateText($text)->passed);
+		$text = "0\n4 2\nUPDATE 2 2 4\nQUERY 1 1 1 3 3 3";
+		$processor = new TextProcessor();
+		$this->assertFalse($processor->ValidateText($text)->passed);
 	}
 
 }
